requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
anthropic==0.34.2
openpyxl==3.1.5
python-multipart==0.0.9
httpx==0.27.2
python-dateutil==2.9.0
tavily-python==0.3.8
